requirements.txt
# Core data
duckdb>=0.10.0
pandas>=2.2.0
pyarrow>=15.0.0

# dbt
dbt-core>=1.8.0
dbt-duckdb>=1.8.0

# API ingestion
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0

# LangChain / AI agent
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0

# FastAPI
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Streamlit dashboard
streamlit>=1.35.0
plotly>=5.22.0
altair>=5.3.0

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.4.0
